Moved configuration exception to it's rightful place
