+++ demos/common/src/main/java/dk/alexandra/fresco/demo/cli/CmdLineProtocolSuite.java
@@ -1,6 +1,5 @@
 package dk.alexandra.fresco.demo.cli;
 
-import dk.alexandra.fresco.framework.configuration.ConfigurationException;
 import dk.alexandra.fresco.framework.sce.resources.ResourcePool;
 import dk.alexandra.fresco.framework.sce.resources.ResourcePoolImpl;
 import dk.alexandra.fresco.framework.sce.resources.storage.FilebasedStreamedStorageImpl;
