+++ demos/common/src/main/java/dk/alexandra/fresco/demo/cli/ConfigurationException.java
@@ -1,4 +1,4 @@
-package dk.alexandra.fresco.framework.configuration;
+package dk.alexandra.fresco.demo.cli;
 
 /**
  * Exception which notifies of FRESCO configuration mistake.
